+++ example/main.tf
@@ -1,3 +1,29 @@
+/*
+=================================================
+
+  Create a 'terrapack' image with Azure-builder
+
+  Note:
+    Variables/locals, data-sources and outputs
+    are included in this file for easier
+    readability. It should be split into
+    separate files
+
+  Tip:
+    It looks cleaner to not provide the
+    module-variables inline, like it is done
+    in this example. See below:
+
+    module some_module_name {
+      source   = "<module-source-url>"
+      version  = "<module-version>"
+      required = local.packer_required_variables
+      optional = local.packer_optional_variables
+    }
+
+=================================================
+*/
+
 variable azure {}
 
 locals {
@@ -10,28 +36,34 @@ resource azurerm_resource_group "PACKER" {
   location = var.azure.resource_group_location
 }
 
-# Build Azure-ARM packer image
+# Build an Azure-ARM packer image
 module packer_example {
-  source     = "../." # You should use registry url
-  #version = "v0.1.0" # You should use pinned version
-  #depends_on = [azurerm_resource_group.PACKER]
-  # --
-  required = { build_id = local.packer_image_name }
+  source   = "../." # You should use registry url w/pinned version
+  required = {
+    build_id = local.packer_image_name # If this value changes, a new build is triggered
+  }
   optional = {
     environment = {
-      # Azure Authentication
-      PKR_VAR_subscription_id  = var.azure.subscription_id
-      PKR_VAR_tenant_id        = var.azure.tenant_id
-      PKR_VAR_object_id        = var.azure.object_id
-      PKR_VAR_client_id        = var.azure.client_id
-      PKR_VAR_client_cert_path = "${path.cwd}/../.credentials/azure-packer.pem"
-
-      # Azure Existing Resources
+      PKR_VAR_subscription_id         = var.azure.subscription_id
+      PKR_VAR_tenant_id               = var.azure.tenant_id
+      PKR_VAR_object_id               = var.azure.object_id
+      PKR_VAR_client_id               = var.azure.client_id
+      PKR_VAR_client_cert_path        = "${path.cwd}/../.credentials/azure-packer.pem"
       PKR_VAR_resource_group_name     = azurerm_resource_group.PACKER.name
       PKR_VAR_resource_group_location = azurerm_resource_group.PACKER.location
-      
-      # Other
-      PKR_VAR_packer_image_name = local.packer_image_name
+      PKR_VAR_packer_image_name       = local.packer_image_name
     }
   }
+}
+
+# Lookup the created image
+data azurerm_image "PACKER_LATEST" {
+  depends_on          = [module.packer_example]
+  resource_group_name = azurerm_resource_group.PACKER.name
+  name                = local.packer_image_name
+}
+
+# Printout created image id
+output image_id {
+  value = data.azurerm_image.PACKER_LATEST.id
 }