+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
+
+## v0.1.1 (Unreleased)
+#### IMPROVEMENTS
+* Example: Improved & cleaned up the included example
+* Packer: Added 'exit 1' on packer failure to default packer-command
+
 ## v0.1.0 (October 08, 2020)
 #### INFO
 * Initial release